Add emulator configuration management, UI, and refactor emulator handling
- Update project version to 1.3.0 in `pom.xml`
- Add `EmuConfigLayoutController` and `EmuConfigLayout.fxml`
- Enhance `LauncherLayoutController` for emulator management
- Implement emulator selection and validation logic
- Add new module definition for emulator configuration
- Improve error handling and user prompts for emulator setup
- Add new keys to `EnumConfigurationKey` for emulator paths
- Change `DTOPoint` fields from double to int
- Implement `Emulator` class with ADB command methods
- Refactor `EmulatorManager` to use new emulator classes and improve
initialization

+++ wos-ot/src/main/java/cl/camodev/wosbot/ot/DTOPoint.java
@@ -1,19 +1,19 @@
 package cl.camodev.wosbot.ot;
 
 public class DTOPoint {
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 
-	public DTOPoint(double x, double y) {
+	public DTOPoint(int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 
-	public double getY() {
+	public int getY() {
 		return y;
 	}
